Extract MockedGameTest into a separate file
